+++ push_swap/libft/ft_memcpy.c
@@ -6,13 +6,13 @@
 /*   By: msamilog <tahasamiloglu@gmail.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/03 14:45:50 by msamilog          #+#    #+#             */
-/*   Updated: 2023/12/01 14:46:50 by msamilog         ###   ########.fr       */
+/*   Updated: 2023/07/03 14:45:50 by msamilog         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memcpy(void *dst, const void *src, int n)
+void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
 	char		*d;
 	const char	*s;
